+++ tests/variable_creation_test.py
@@ -372,3 +372,71 @@ def test_empty_sizes():
                         sc.empty(sizes=dict(zip(dims, shape))))
     with pytest.raises(ValueError):
         sc.empty(dims=dims, shape=shape, sizes=dict(zip(dims, shape)))
+
+
+def test_datetime():
+    assert sc.identical(sc.datetime('1970', unit='Y'),
+                        sc.scalar(np.datetime64('1970', 'Y')))
+    assert sc.identical(sc.datetime('2015-06-13'),
+                        sc.scalar(np.datetime64('2015-06-13', 'D')))
+    assert sc.identical(sc.datetime('2152-11-25T13:13:46'),
+                        sc.scalar(np.datetime64('2152-11-25T13:13:46', 's')))
+    assert sc.identical(sc.datetime('2152-11-25T13:13:46', unit='h'),
+                        sc.scalar(np.datetime64('2152-11-25T13', 'h')))
+    assert sc.identical(sc.datetime('2152-11-25T13:13:46', unit='us'),
+                        sc.scalar(np.datetime64('2152-11-25T13:13:46', 'us')))
+
+    assert sc.identical(sc.datetime(626, unit='s'),
+                        sc.scalar(626, dtype='datetime64', unit='s'))
+    assert sc.identical(sc.datetime(2**10, unit='ns'),
+                        sc.scalar(2**10, dtype='datetime64', unit='ns'))
+    assert sc.identical(sc.datetime(-94716, unit='min'),
+                        sc.scalar(-94716, dtype='datetime64', unit='min'))
+
+    assert sc.identical(sc.datetime(np.datetime64(314, 's'), unit='s'),
+                        sc.scalar(314, dtype='datetime64', unit='s'))
+    assert sc.identical(sc.datetime(np.datetime64(671, 'h')),
+                        sc.scalar(671, dtype='datetime64', unit='h'))
+
+
+def test_datetimes():
+    assert sc.identical(
+        sc.datetimes(dims=['t'], values=['1970', '2021'], unit='Y'),
+        sc.array(dims=['t'],
+                 values=[np.datetime64('1970', 'Y'),
+                         np.datetime64('2021', 'Y')],
+                 unit='Y'))
+    assert sc.identical(
+        sc.datetimes(dims=['t'],
+                     values=['2152-11-25T13:13:46', '1111-11-11T11:11:11'],
+                     unit='s'),
+        sc.array(dims=['t'],
+                 values=[
+                     np.datetime64('2152-11-25T13:13:46', 's'),
+                     np.datetime64('1111-11-11T11:11:11', 's')
+                 ],
+                 unit='s'))
+    assert sc.identical(
+        sc.datetimes(dims=['t'],
+                     values=['2152-11-25T13:13:46', '1111-11-11T11:11:11'],
+                     unit='us'),
+        sc.array(dims=['t'],
+                 values=[
+                     np.datetime64('2152-11-25T13:13:46', 'us'),
+                     np.datetime64('1111-11-11T11:11:11', 'us')
+                 ],
+                 unit='us'))
+
+    assert sc.identical(
+        sc.datetimes(dims=['t'], values=[0, 123, 2**10], unit='s'),
+        sc.array(dims=['t'], values=np.array([0, 123, 2**10], dtype='datetime64[s]')))
+    assert sc.identical(
+        sc.datetimes(dims=['t'], values=[-723, 2**13, -3**5], unit='min'),
+        sc.array(dims=['t'],
+                 values=np.array([-723, 2**13, -3**5], dtype='datetime64[m]')))
+
+
+def test_datetime_epoch():
+    assert sc.identical(sc.epoch(unit='s'),
+                        sc.scalar(np.datetime64('1970-01-01T00:00:00', 's')))
+    assert sc.identical(sc.epoch(unit='D'), sc.scalar(np.datetime64('1970-01-01', 'D')))